requirements/pip_freeze.txt
aniso8601==7.0.0
Click==7.0
Flask==1.1.0
Flask-JWT-Extended==3.20.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
gunicorn==19.9.0
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
passlib==1.7.1
psycopg2-binary==2.8.3
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.5
SQLAlchemy-Utils==0.34.0
SQLAlchemy-Wrapper==1.9.1
Werkzeug==0.15.4
